backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pytest==7.3.0
pytest-cov==4.1.0
python-dotenv==1.0.0
